requirements.txt
appdirs==1.4.3
Django==1.11
PyJWT==1.4.2
djangorestframework==3.6.2
djangorestframework-jwt==1.8.0
packaging==16.8
pyparsing==2.2.0
pytz==2017.2
six==1.10.0
wheel==0.29.0
gunicorn
whitenoise
MySQL-python==1.2.5